Add files via upload
the two undulating rigid NACA0012 hydrofoils are
fixed at their initial locations, i.e. the foils are ‘‘tethered’’. This tethered
configuration to study linearly accelerating swimmers has been
justified by the experimental study of Akanyeti et al. (2017), which
also compared the experimental data to biological statistics. To account
for the accelerating condition, the mean net thrust on the swimmers
is not zero. It should be noted that the present setup differs from
many previous fish schooling studies focusing on the steady swimming
condition with zero net force. The non-dimensional form of the physical
problem investigated in the present paper is demonstrated in Fig. 2.
Here, 𝐶 = 1 is the chord length of the two NACA0012 hydrofoils.
𝐺 and 𝐷 are the lateral and front–back distances between the two
hydrofoils. The computational domain is 16𝐶 in the streamwise direction
and 8𝐶 in  We note
that we are not strictly modelling any specific fish species but are
instead seeking underlying principles of how swimming style affects
accelerated schooling.
In addition, similar problem setups have been adopted and justified
by robotic experiments (Akanyeti et al., 2017) and numerical simulation
(Thekkethil et al., 2017, 2018, 2020; Gupta et al., 2021) to study
accelerating swimmers.